Get rid of most server code for providers
I want to drop the provider tables, since they're unused and there's a
lot of crap around them and I think it'll be confusing for people
dealing with the schema anew to have all these concepts running around
that aren't necessary.

This diff unceremoneously bins everything related to providers from the
admin tool, including GraphQL mutations and such (no need for backwards
compatibility here), along with the `defaultProvider` field on
`ApplicationEntity` and a couple places related to that and the
`providersForDomain` GraphQL query that hasn't been called in the last
7d per go/deprecations.

Test Plan:

Everything compiles, and I can still save changes to applications in the
admin tool.