Remove the providerID column from PageEntity
This column is going to need to go away when we drop the provider
tables, so remove it from the entity in preparation for that.

What will ultimately happen here is we're going to just drop the
`providerID` column from the table. That will end up basically
abandoning all the rows that previously had a `providerID` set, because
recomputing the context hash for those rows would produce a different
value, so anything that would try to retrieve those rows now looks in a
different context hash. This doesn't end up causing a problem, because
we never refer to rows by their context data from other tables, we link
to them via the context hash. But that's already what happens in the
SDK: the page hash computing function doesn't even consider providerID
anymore. So this is safe (and those rows are also abandoned already).

This also doesn't affect any data that anyone cares about. The last
thread that is linked to a row with a `providerID` set is from April
2023, and the last one that's not in the `cord-extension` application is
from October 2022, for threads from the embed distribution (which
doesn't exist anymore).

So basically we're going to abandon a bunch of rows that nobody can
access anymore anyway.

Test Plan:

Send a message in a location that didn't previously exist, it was
created properly.