Add endpoint to dump DB data for a customer
This adds an endpoint that dumps the customer's DB data so that it can
be imported into Postgres.

This works for everything except Slack users. We don't want to export
the whole list of Slack users/orgs, since there's a lot of stuff there
and those users may have linked to a different customer's application
rather than this customer's. However, we do have to export them at some
point, since they can have authored messages and such, so the data
export doesn't work without them.

Right now, we just skip the `linked_users` and `linked_orgs` tables and
require that users have an appropriate applicationID. This won't
actually work yet (we'll have messages that try to import with a
sourceID that points to a user that doesn't exist), but it's a start.

Test Plan:

Ran

```
curl https://local.cord.com:8161/v1/customer/dbdump --oauth2-bearer $TOKEN
```

and inspected the results. The SQL queries look good (albeit ugly in
some cases), and the actual rows seem sensible.