Add handling of Slack users to DB dump
Add support for Slack users to the DB dump handler, as well as a few
more things to make it actually work in practice. The Slack handling is
done by exporting any user that either has authored a message or shows
up in the `linked_users` table, and Slack orgs are exported if they show
up in the `linked_orgs` table.

Order the tables to make sure that foreign key references come first,
before the tables with their keys. That means `customers` before
`applications`, `applications` before `users`, `threads` before
`messages`, etc.

Add a `DEFERRABLE INITIALLY DEFERRED` to the support bot columns,
because there's a circular reference between bots and their applications
and we need to break it.

This isn't yet the final version, because the query to get foreign keys
is broken (sometimes it mismatches the columns for multi-column foreign
keys), which means sometimes it misses data to export. Leaving that for
a future diff since it's a localized fix to the foreign key-finding
query.

Test Plan:

Grabbed a dump and tried to load it into an empty DB, it loaded
successfully.