Speed up/correctify dumping notifications table
The query for the notifications table was very complicated because it
has foreign keys to a half dozen different tables, some portion of which
are set at any given time, and so the automatic generation system was
really complicated. However, it also just has a `platformApplicationID`
column, so skip autogenerating everything and just go direct to make
sure it belongs to the right application. It was also missing some rows
because of the autogenerated stuff, so that's fixed as a side effect.

This also revealed that we have notifications that have a recipientID of
a Slack user that doesn't show up in either of the other places we were
finding them, so include them as well.

Test Plan:

Created a dump and loaded it into an empty DB, this reduced the time to
dump from about 15m to 2m.