Add S3 file copying handler
Add a new handler that copies files from S3 to a destination bucket.

The customer will need to give our production user put + list
permissions to the bucket they want to transfer files to. Then they call
this handler with the region + bucket that they want to copy their files
to. The handler looks up the files for that customer's applications and
asks S3 to internally copy it to the destination bucket.

The file copier checks the filenames in the destination bucket and skips
copying anything that's already in the bucket, so this can be used to
incrementally copy all the files over, as well as to do it over time so
that downtime for an eventual cutover is minimized.

Test Plan:

Changed my local .env to use the production S3 bucket, then uploaded a
file and requested that the handler copy it to a temp bucket I created.
It worked properly the first time I called it, then copied nothing the
second time I called it.