Change file copy handler to work in parallel
The file handler is a little slow if you're trying to copy 1000 files at
a time. Do all the CopyObject calls in parallel, since S3 doesn't care.

We still do the first one not in parallel, because our user not having
permission to write to the dest bucket is probably pretty likely, and
there's no point in firing off 1000 requests that all are going to fail.
Once the first one succeeds, though, we can be pretty sure all the rest
are likely to.

Test Plan:

Ran it locally to copy set of files, it worked.