Fix cross-region file copies
It turns out if you want to copy an S3 file cross-region, you have to
send the request to the destination region rather than the source
region. That is very surprising to me, but it seems to work.

Test Plan:

Created a bucket in another region and requested copying files to it, it
failed before this diff and worked after it.